Rename KV Service API Pacakge
