+++ pkg/service/kv-api/delete.go
