+++ server/main.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 	"time"
 
-	service "github.com/RyanDerr/GoKeyValueStore/pkg/service/cache-api"
+	service "github.com/RyanDerr/GoKeyValueStore/pkg/service/kv-api"
 	pb "github.com/RyanDerr/GoKeyValueStore/proto-public/go"
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"google.golang.org/grpc"
